Split learn documentation parser into separated parser per page
